Narrowed down Italic Options
Narrowed down to two Italic options and replicated those for the
different optical sizes. All other files should most mostly be the same
as the previous dated folder (G Serif 170522). There are various web
proofs made to compare and contrast the different italics at different
sizes in proofs folder. There is also a folder of killed italics we
felt weren’t working for the screen and in matching the romans.