Italic options being narrowed even more
After narrowing down to two italic options for each optical size from
the previous push (G Serif 170707), we felt that italic option 02
worked better for the display while italic option 01 worked better for
the micro because it was more structured and is wider which helps with
readability at smaller sizes. We decided to interpolate between the two
italic options to get an idea of what the middle would look like,
especially if we were going to have italic 01 for micro sizes and 02
for display. Most of the changes to these UFOs are to make them
interpolatable.